consolidate flow of arrows attack
