+++ SolastaUnfinishedBusiness/CustomBehaviors/AddExtraAttack.cs
@@ -82,16 +82,6 @@ internal abstract class AddExtraAttackBase : IAddExtraAttack
         }
     }
 
-#if false
-    [NotNull]
-    internal AddExtraAttackBase SetTags([NotNull] params string[] tags)
-    {
-        additionalTags.AddRange(tags);
-
-        return this;
-    }
-#endif
-
     protected abstract List<RulesetAttackMode> GetAttackModes(RulesetCharacter character);
 
     protected virtual AttackModeOrder GetOrder(RulesetCharacter character)
@@ -217,7 +207,9 @@ internal sealed class AddExtraMainHandAttack : AddExtraAttackBase
         // don't use ?? on Unity Objects as it bypasses the lifetime check on the underlying object
         var strikeDefinition = mainHandItem?.ItemDefinition;
 
+#pragma warning disable IDE0270
         if (strikeDefinition == null)
+#pragma warning restore IDE0270
         {
             strikeDefinition = hero.UnarmedStrikeDefinition;
         }
@@ -322,10 +314,6 @@ internal sealed class AddPolearmFollowUpAttack : AddExtraAttackBase
         var result = new List<RulesetAttackMode>();
 
         AddItemAttack(result, EquipmentDefinitions.SlotTypeMainHand, hero);
-
-        // doesn't make sense to add a bonus attack from an offhand slot that already uses your bonus action
-
-        // AddItemAttack(result, EquipmentDefinitions.SlotTypeOffHand, hero);
 
         return result;
     }
@@ -508,7 +496,6 @@ internal sealed class AddBonusTorchAttack : AddExtraAttackBase
         attackMode.Ranged = false;
         attackMode.Thrown = false;
         attackMode.AutomaticHit = true;
-
         attackMode.EffectDescription.Clear();
         attackMode.EffectDescription.Copy(_torchPower.EffectDescription);
 
@@ -519,5 +506,43 @@ internal sealed class AddBonusTorchAttack : AddExtraAttackBase
             8 + proficiencyBonus + AttributeDefinitions.ComputeAbilityScoreModifier(dexterity);
 
         attackModes.Add(attackMode);
+    }
+}
+
+internal sealed class AddExtraFlurryOfArrowsAttack : AddExtraAttackBase
+{
+    public AddExtraFlurryOfArrowsAttack() :
+        base(ActionDefinitions.ActionType.Bonus, ValidatorsCharacter.HasBowWithoutArmor)
+    {
+    }
+
+    protected override AttackModeOrder GetOrder(RulesetCharacter character)
+    {
+        return AttackModeOrder.Start;
+    }
+
+    protected override List<RulesetAttackMode> GetAttackModes(RulesetCharacter character)
+    {
+        if (character is not RulesetCharacterHero hero || !ValidatorsCharacter.HasBowWithoutArmor(hero))
+        {
+            return null;
+        }
+
+        var mainHandItem = hero.GetMainWeapon();
+        var attackModifiers = hero.attackModifiers;
+        var attackMode = hero.RefreshAttackMode(
+            ActionType,
+            mainHandItem!.ItemDefinition,
+            mainHandItem.ItemDefinition.WeaponDescription,
+            false,
+            true,
+            EquipmentDefinitions.SlotTypeMainHand,
+            attackModifiers,
+            hero.FeaturesOrigin,
+            mainHandItem);
+
+        attackMode.attacksNumber = 1;
+
+        return new List<RulesetAttackMode> { attackMode };
     }
 }
